requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Data processing and export
pandas>=1.5.0
openpyxl>=3.0.0

# Web scraping utilities
selenium>=4.0.0
urllib3>=1.26.0

# Email functionality (optional)
# smtplib is built-in

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Async support (optional)
aiohttp>=3.8.0

# Rate limiting
ratelimit>=2.2.0
